+++ SysBot.Pokemon.Slack/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace SysBot.Pokemon.Slack;
+
+public class Class1
+{
+
+}